Rename openai.py to openai_wrapper.py
Turns out super weird edge case where you can't name any files `openai.py`: https://community.openai.com/t/attributeerror-module-openai-has-no-attribute-chatcompletion/81490
