+++ src/tracing_auto_instrumentation/openai/__init__.py
@@ -1,4 +1,4 @@
-from .openai import wrap_openai
+from .openai_wrapper import wrap_openai
 
 __ALL__ = [
     wrap_openai.__name__,
